requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
docx2pdf==0.1.8
Flask==3.1.1
img2pdf==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
pikepdf==9.7.0
pillow==11.2.1
python-docx==1.1.2
tqdm==4.67.1
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
fpdf==1.7.2
pdf2image==1.16.3
pdf2docx==0.5.6
Pillow==11.2.0
